// src/shared/ui/flex/model/types/flex-type.ts
// types
export type FlexJustifyContent =
  | "center"
  | "end"
  | "spaceAround"
  | "spaceBetween"
  | "spaceEvenly"
  | "start";

export type FlexGapType =
  | "1"
  | "2"
  | "3"
  | "4"
  | "6"
  | "8"
  | "10"
  | "12"
  | "14"
  | "16"
  | "18"
  | "20"
  | "22"
  | "24"
  | "26"
  | "28"
  | "30"
  | "34"
  | "40"
  | "50"
  | "60"
  | "70"
  | "80";

export type FlexAlignItemsType =
  | "baseline"
  | "center"
  | "end"
  | "start"
  | "stretch";
export type FlexDirectionType =
  | "col_reverse"
  | "column"
  | "row"
  | "row_reverse";
export type FlexWrapType = "no_wrap" | "wrap" | "wrap_reverse";